Fix infinite loop build
When running the npm watch command, the building process is locked in an
infinite loop. It is apparently due to the fact that webpack doesn’t
support the glob option when registering PostCSS dependencies.

In order to fix it we can tell webpack to only watch specific files or
in my case use the fast-glob npm package.

See this link for a better explanation of the problem:
https://tailwindcss.com/docs/content-configuration#styles-rebuild-in-an-infinite-loop

Read article about how to solve this with fast-glob: https://tips.dev/tailwind-css/how-to-fix-laravel-mix-and-tailwind-infinite-loop/